Moved functions to parent class ProviderWidget
